octocrab

Module auth

Authentication related types and functions.

Structs

AppAuth

The data necessary to authenticate as a Github App

DeviceCodes

The device codes as returned from step 1 of Github’s device flow.

OAuth

The data necessary to authenticate as a GitHub OAuth app.

Enums

Auth

The forms of authentication we support

Continue

Control flow when polling the device flow authorization.

Functions

create_jwt

Create a JSON Web Token that can be used to authenticate an a GitHub application.