octocrab
Module current

Get data about the currently authenticated user.

Structs

CurrentAuthHandler

Handler for the current authenication API. Note All of the methods provided below require at least some authenication such as personal token in order to be used.

ListAppInstallationsAccessibleToUserBuilder

A builder pattern struct for listing the installations accessible to a user access token.

ListGistsForAuthenticatedUserBuilder

A builder struct for initializing query parameters for use with the /gists endpoint.

ListOrgMembershipsForAuthenticatedUserBuilder

A builder pattern struct for listing organizations the authenticated user is a member of.

ListReposForAuthenticatedUserBuilder

A builder pattern struct for listing repositories for authenticated user.

ListStarredGistsBuilder

ListStarredReposBuilder

A builder pattern struct for listing starred repositories.