octocrab::etag

Struct EntityTag

pub struct EntityTag {
    pub weak: bool,
    /* private fields */
}

An entity tag, defined in RFC7232

An entity tag consists of a string enclosed by two literal double quotes. Preceding the first double quote is an optional weakness indicator, which always looks like W/. Examples for valid tags are "xyzzy" and W/"xyzzy".

ABNF

entity-tag = [ weak ] opaque-tag
weak       = %x57.2F ; "W/", case-sensitive
opaque-tag = DQUOTE *etagc DQUOTE
etagc      = %x21 / %x23-7E / obs-text
           ; VCHAR except double quotes, plus obs-text

Comparison

To check if two entity tags are equivalent in an application always use the strong_eq or weak_eq methods based on the context of the Tag. Only use == to check if two tags are identical.

The example below shows the results for a set of entity-tag pairs and both the weak and strong comparison function results:

ETag 1	ETag 2	Strong Comparison	Weak Comparison

W/"1"	W/"1"	no match	match

W/"1"	W/"2"	no match	no match

W/"1"	"1"	no match	match

"1"	"1"	match	match

Fields
weak: bool

Weakness indicator for the tag

Implementations

impl EntityTag

pub fn extract_from_response<B>(response: &Response<B>) -> Option<EntityTag>

pub fn insert_if_none_match_header( headers: &mut HeaderMap, etag: EntityTag, ) -> Result<(), Error>

pub fn new(weak: bool, tag: String) -> EntityTag

Constructs a new EntityTag.

Panics

If the tag contains invalid characters.

pub fn weak(tag: String) -> EntityTag

Constructs a new weak EntityTag.

Panics

If the tag contains invalid characters.

pub fn strong(tag: String) -> EntityTag

Constructs a new strong EntityTag.

Panics

If the tag contains invalid characters.

pub fn tag(&self) -> &str

Get the tag.

pub fn set_tag(&mut self, tag: String)

Set the tag.

Panics

If the tag contains invalid characters.

pub fn strong_eq(&self, other: &EntityTag) -> bool

For strong comparison two entity-tags are equivalent if both are not weak and their opaque-tags match character-by-character.

pub fn weak_eq(&self, other: &EntityTag) -> bool

For weak comparison two entity-tags are equivalent if their opaque-tags match character-by-character, regardless of either or both being tagged as โweakโ.

pub fn strong_ne(&self, other: &EntityTag) -> bool

The inverse of EntityTag.strong_eq().

pub fn weak_ne(&self, other: &EntityTag) -> bool

The inverse of EntityTag.weak_eq().