octocrab

Module models

Serde mappings from GitHub’s JSON to structs.

Re-exports

pub use apps::App;

Modules

actions

activity

apps

checks

code_scannings

commits

events

gists

hooks

issues

orgs

pulls

reactions

repos

teams

timelines

webhook_events

Serde mappings from GitHub’s Webhook payloads to structs.

workflows

Structs

ActorId

AppId

ArtifactId

AssetId

Author

BranchProtectionRuleId

CardId

CheckRun

CheckRunId

CheckRuns

CheckSuiteId

Code

CodeScanningId

Collaborator

CombinedStatus

CommentId

Contents

Contributor

Followee

A user that is being followed by another user

Follower

A user that is following another user

GitSshKey

GpgKey

HookDeliveryId

HookId

Installation

InstallationId

InstallationPermissions

InstallationRepositories

InstallationToken

IssueEvent

IssueEventId

IssueId

IssuePullRequest

JobId

Label

LabelId

License

Milestone

MilestoneId

NotificationId

OrgId

Permissions

Project

ProjectCard

ProjectColumn

ProjectColumnId

ProjectId

PublicKey

PullRequestId

PushId

Rate

RateLimit

ReactionId

ReleaseId

Repository

RepositoryFile

RepositoryId

RepositoryMetrics

Resources

ReviewId

RunId

RunnerGroupId

RunnerId

RunnerLabelId

SimpleUser

The simple profile for a GitHub user

SocialAccount

SshSigningKey

StarGazer

Status

StatusId

SubKeyInfo

TeamId

TeamInvitationId

ThreadId

TimelineEventId

UploaderId

UserEmailInfo

UserId

UserOrOrgId

UserProfile

The full profile for a user

VerifiedEmailInfo

WorkflowId

Enums

AuthorAssociation

CheckStatus

Event

Issue events are triggered by activity in issues and pull requests. https://docs.github.com/en/webhooks-and-events/events/issue-event-types

IssueState

ProjectCardContentType

StatusState