octocrab::models::events
Module payload

Structs

Commit

A git commit in specific payload types.

CommitCommentEventPayload

The payload in a super::EventPayload::CommitCommentEvent type.

CreateEventPayload

The payload in a super::EventPayload::CreateEvent type.

DeleteEventPayload

The payload in a super::EventPayload::DeleteEvent type.

EventInstallationId

ForkEventPayload

The payload in a super::EventPayload::ForkEvent type.

GollumEventPage

A page in a GollumEventPayload.

GollumEventPayload

The payload in a super::EventPayload::GollumEvent type.

InstallationEventPayload

The payload in a webhook installation event type.

InstallationEventRepository

A repository in installation related webhook events.

InstallationRepositoriesEventPayload

The payload in a webhook installation_repositories event type.

InstallationTargetChanges

InstallationTargetEventPayload

The payload in a webhook installation_target event type.

InstallationTargetLoginChanges

InstallationTargetSlugChanges

IssueCommentEventChangesFrom

The previous value of the body of the issue comment which has changed. Only available in an event of type IssueCommentEventAction::Edited.

IssueCommentEventPayload

The payload in a super::EventPayload::IssueCommentEvent type.

IssuesEventChangesFrom

The previous value of the item (either the body or title) of an issue which has changed. Only available in an event of type IssuesEventAction::Edited.

IssuesEventPayload

The payload in a super::EventPayload::IssuesEvent type.

MemberEventChangesTo

The new permission given to the collaborator.

MemberEventPayload

The payload in a super::EventPayload::MemberEvent type.

PublicEventPayload

The payload in a super::EventPayload::PublicEvent type.

PullRequestChanges

The change which occurred in an event of type PullRequestEventAction::Edited.

PullRequestEventChangesFrom

The previous value of the item (either the body or title) of a pull request which has changed. Only available in an event of type PullRequestEventAction::Edited.

PullRequestEventPayload

The payload in a super::EventPayload::PullRequestEvent type.

PullRequestReviewChangesFrom

The previous value of the body of a review which has changed. Only available in an event of type [PullRequestReviewEventAction::Edited].

PullRequestReviewCommentChangesFrom

The previous value of the body of a review comment which has changed. Only available in an event of type PullRequestReviewCommentEventAction::Edited.

PullRequestReviewCommentEventPayload

The payload in a super::EventPayload::PullRequestReviewCommentEvent type.

PullRequestReviewEventPayload

The payload in a super::EventPayload::PullRequestReviewEvent type.

PushEventPayload

The payload in a super::EventPayload::PushEvent type.

ReleaseEventChangesFrom

The previous value of the item (either the body or title) of a release which has changed. Only available in an event of type ReleaseEventAction::Edited.

ReleaseEventPayload

The payload in a super::EventPayload::ReleaseEvent type.

WatchEventPayload

The payload in a super::EventPayload::WatchEvent type.

WorkflowRunEventPayload

The payload in a super::EventPayload::WorkflowRunEvent type.

WrappedEventPayload

Enums

EventInstallation

EventPayload

The payload in an event type.

GollumEventPageAction

The action performed on a given page.

InstallationEventAction

The action on an installation this event corresponds to.

InstallationRepositoriesEventAction

The action on an installation this event corresponds to.

InstallationRepositoriesEventSelection

IssueCommentEventAction

The action on an issue comment this event corresponds to.

IssueCommentEventChanges

The change which occurred in an event of type IssueCommentEventAction::Edited.

IssuesEventAction

The action on an issue this event corresponds to.

IssuesEventChanges

The change which occurred in an event of type IssuesEventAction::Edited.

MemberEventAction

The action this event corresponds to.

MemberEventChanges

The change which occurred.

PullRequestEventAction

The action on a pull request this event corresponds to.

PullRequestReviewChanges

The change which occurred in an event of type [PullRequestReviewEventAction::Edited].

PullRequestReviewCommentChanges

The change which occurred in an event of type PullRequestReviewCommentEventAction::Edited.

PullRequestReviewCommentEventAction

The action on a pull request review comment this event corresponds to.

PullRequestReviewEventAction

The action on a pull request review this event corresponds to.

ReleaseEventAction

The action on a release this event corresponds to.

ReleaseEventChanges

The change which occurred in an event of type ReleaseEventAction::Edited.

WatchEventAction

The action that was performed as part of the super::EventPayload::WatchEvent.

WorkflowRunEventAction

The action on a pull request this event corresponds to.