Struct octocrab::repos::ReleasesHandler

pub struct ReleasesHandler<'octo, 'r> { /* private fields */ }

Handler for GitHubโs releases API.

Created with RepoHandler::releases.

Implementations

impl<'octo, 'r> ReleasesHandler<'octo, 'r>

pub fn list(&self) -> ListReleasesBuilder<'_, '_, '_>

Creates a new ListReleasesBuilder that can be configured to filter listing releases.

let page = octocrab.repos("owner", "repo")
    .releases()
    .list()
    // Optional Parameters
    .per_page(100)
    .page(5u32)
    // Send the request
    .send()
    .await?;

pub fn create<'t>( &self, tag_name: &'t (impl AsRef<str> + ?Sized), ) -> CreateReleaseBuilder<'_, '_, '_, 't, '_, '_, '_>

Creates a new CreateReleaseBuilder with tag_name.

let page = octocrab.repos("owner", "repo")
    .releases()
    .create("v1.0.0")
    // Optional Parameters
    .target_commitish("main")
    .name("Version 1.0.0")
    .body("Announcing 1.0.0!")
    .draft(false)
    .prerelease(false)
    // Send the request
    .send()
    .await?;

pub fn update( &self, release_id: u64, ) -> UpdateReleaseBuilder<'_, '_, '_, '_, '_, '_, '_>

Creates a new UpdateReleaseBuilder with release_id.

let release = octocrab.repos("owner", "repo")
    .releases()
    .update(1)
    // Optional Parameters
    .tag_name("v1.0.0")
    .target_commitish("main")
    .name("Version 1.0.0")
    .body("Announcing 1.0.0!")
    .draft(false)
    .prerelease(false)
    // Send the request
    .send()
    .await?;

pub async fn get_asset(&self, asset_id: u64) -> Result<Asset>

๐Deprecated: use repos::ReleaseAssetsHandler::get instead

Fetches a single asset by its ID.

let asset = octocrab::instance()
    .repos("owner", "repo")
    .releases()
    .get_asset(42u64.into())
    .await?;

pub async fn get_latest(&self) -> Result<Release>

Gets the latest release.

let release = octocrab::instance()
    .repos("owner", "repo")
    .releases()
    .get_latest()
    .await?;

pub async fn get_by_tag(&self, tag: &str) -> Result<Release>

Gets the release using its tag.

let release = octocrab::instance()
    .repos("owner", "repo")
    .releases()
    .get_by_tag("v1.0.0")
    .await?;

pub async fn get(&self, number: u64) -> Result<Release>

Gets the release using its id.

let release = octocrab::instance()
    .repos("owner", "repo")
    .releases()
    .get(3)
    .await?;

pub fn generate_release_notes<'tag_name>( &self, tag_name: &'tag_name (impl AsRef<str> + ?Sized), ) -> GenerateReleaseNotesBuilder<'_, '_, '_, 'tag_name, '_, '_, '_>

Generates crate::models::repos::ReleaseNotes which describe a crate::models::repos::Release

let release_notes = octocrab::instance()
    .repos("owner", "repo")
    .releases()
    .generate_release_notes("0.1.0")
    .send()
    .await?;

pub fn upload_asset<'asset_name>( &self, id: u64, asset_name: &'asset_name (impl AsRef<str> + ?Sized), body: Bytes, ) -> UploadAssetBuilder<'_, '_, '_, 'asset_name, '_>

Upload an crate::models::repos::Asset associated with a crate::models::repos::Release

use bytes::Bytes;
let file_data: Bytes = Bytes::from("some_data");
let asset = octocrab::instance()
    .repos("owner", "repo")
    .releases()
    .upload_asset(1, "my_asset.tar.gz", file_data)
    .label("My Awesome Asset")
    .send()
    .await?;

pub fn assets(&self, release_id: u64) -> ListReleaseAssetsBuilder<'_, '_, '_>

Creates a new ListReleaseAssetsBuilder that can be configured to filter listing release assetss.

let page = octocrab.repos("owner", "repo")
    .releases()
    .assets(1)
    // Optional Parameters
    .per_page(100)
    .page(5u32)
    // Send the request
    .send()
    .await?;

pub async fn delete(&self, id: u64) -> Result<()>

Delete a release using its id.

let release = octocrab::instance()
    .repos("owner", "repo")
    .releases()
    .delete(3)
    .await?;