octocrab::service::middleware::cache

Trait CacheStorage

pub trait CacheStorage: Send + Sync {
    // Required methods
    fn try_hit(&self, uri: &Uri) -> Option<CacheKey>;
    fn load(&self, uri: &Uri) -> Option<CachedResponse>;
    fn writer(
        &self,
        uri: &Uri,
        key: CacheKey,
        headers: HeaderMap,
    ) -> Box<dyn CacheWriter>;
}

HttpCacheLayer is agnostic to the storage implementation (e.g., in-memory, filesystem, etc.). This trait represents the required interface.

Required Methods

fn try_hit(&self, uri: &Uri) -> Option<CacheKey>

Returns the stored cache key for given URI if it’s available in the cache.

fn load(&self, uri: &Uri) -> Option<CachedResponse>

Returns the cached response for given URI if it’s available in the cache.

This method is expected to return Some if CacheStorage::try_hit returned Some for given URI.

fn writer( &self, uri: &Uri, key: CacheKey, headers: HeaderMap, ) -> Box<dyn CacheWriter>

Returns a writer that writes the response body to the cache.