Struct octocrab::Octocrab

pub struct Octocrab { /* private fields */ }

The GitHub API client.

Implementations

impl Octocrab

pub async fn authenticate_as_device<I, S>( &self, client_id: &SecretString, scope: I, ) -> Result<DeviceCodes>
where I: IntoIterator<Item = S>, S: AsRef<str>,

Authenticate with Github’s device flow. This starts the process to obtain a new OAuth.

See https://docs.github.com/en/developers/apps/building-oauth-apps/authorizing-oauth-apps#device-flow for details.

Note: To authenticate against public Github, the Octocrab that calls this method must be constructed with base_uri: "https://github.com" and extra header “ACCEPT: application/json”. For example:

let crab = octocrab::Octocrab::builder()
.base_uri("https://github.com")?
.add_header(ACCEPT, "application/json".to_string())
.build()?;

impl Octocrab

Constructors

pub fn builder() -> OctocrabBuilder<NoSvc, DefaultOctocrabBuilderConfig, NoAuth, NotLayerReady>

Returns a new OctocrabBuilder.

pub fn installation(&self, id: InstallationId) -> Result<Octocrab>

Returns a new Octocrab based on the current builder but authorizing via a specific installation ID. Typically you will first construct an Octocrab using OctocrabBuilder::app to authenticate as your Github App, then obtain an installation ID, and then pass that here to obtain a new Octocrab with which you can make API calls with the permissions of that installation.

pub async fn installation_and_token( &self, id: InstallationId, ) -> Result<(Octocrab, SecretString)>

Similar to installation, but also eagerly caches the installation token and returns the token. The returned token can be used to make https git requests to e.g. clone repositories that the installation has access to.

See also https://docs.github.com/en/developers/apps/building-github-apps/authenticating-with-github-apps#http-based-git-access-by-an-installation

impl Octocrab

GitHub API Methods

pub fn actions(&self) -> ActionsHandler<'_>

Creates a new actions::ActionsHandler for accessing information from GitHub Actions.

pub fn current(&self) -> CurrentAuthHandler<'_>

Creates a current::CurrentAuthHandler that allows you to access information about the current authenticated user.

pub fn activity(&self) -> ActivityHandler<'_>

Creates a activity::ActivityHandler for the current authenticated user.

pub fn apps(&self) -> AppsRequestHandler<'_>

Creates a new apps::AppsRequestHandler for the currently authenticated app.

pub fn gitignore(&self) -> GitignoreHandler<'_>

Creates a gitignore::GitignoreHandler for accessing information about gitignore.

pub fn issues( &self, owner: impl Into<String>, repo: impl Into<String>, ) -> IssueHandler<'_>

Creates a issues::IssueHandler for the repo specified at owner/repo, that allows you to access GitHub’s issues API.

pub fn issues_by_id(&self, id: impl Into<RepositoryId>) -> IssueHandler<'_>

Creates a issues::IssueHandler for the repo specified at repository ID, that allows you to access GitHub’s issues API.

pub fn code_scannings( &self, owner: impl Into<String>, repo: impl Into<String>, ) -> CodeScanningHandler<'_>

Creates a [code_scanning::CodeSCanningHandler] for the repo specified at owner/repo, that allows you to access GitHub’s Code scanning API.

pub fn code_scannings_organisation( &self, owner: impl Into<String>, ) -> CodeScanningHandler<'_>

Creates a [code_scanning::CodeSCanningHandler] for the org specified at owner, that allows you to access GitHub’s Code scanning API.

pub fn commits( &self, owner: impl Into<String>, repo: impl Into<String>, ) -> CommitHandler<'_>

Creates a commits::CommitHandler for the repo specified at owner/repo,

pub fn licenses(&self) -> LicenseHandler<'_>

Creates a licenses::LicenseHandler.

pub fn markdown(&self) -> MarkdownHandler<'_>

Creates a markdown::MarkdownHandler.

pub fn orgs(&self, owner: impl Into<String>) -> OrgHandler<'_>

Creates an orgs::OrgHandler for the specified organization, that allows you to access GitHub’s organization API.

pub fn pulls( &self, owner: impl Into<String>, repo: impl Into<String>, ) -> PullRequestHandler<'_>

Creates a pulls::PullRequestHandler for the repo specified at owner/repo, that allows you to access GitHub’s pull request API.

pub fn repos( &self, owner: impl Into<String>, repo: impl Into<String>, ) -> RepoHandler<'_>

Creates a repos::RepoHandler for the repo specified at owner/repo, that allows you to access GitHub’s repository API.

pub fn repos_by_id(&self, id: impl Into<RepositoryId>) -> RepoHandler<'_>

Creates a repos::RepoHandler for the repo specified at repository ID, that allows you to access GitHub’s repository API.

pub fn projects(&self) -> ProjectHandler<'_>

Creates a projects::ProjectHandler that allows you to access GitHub’s projects API (classic).

pub fn search(&self) -> SearchHandler<'_>

Creates a search::SearchHandler that allows you to construct general queries to GitHub’s API.

pub fn teams(&self, owner: impl Into<String>) -> TeamHandler<'_>

Creates a teams::TeamHandler for the specified organization that allows you to access GitHub’s teams API.

pub fn users(&self, user: impl Into<String>) -> UserHandler<'_>

Creates a [users::UserHandler] for the specified user using the user name

pub fn users_by_id(&self, user: impl Into<UserId>) -> UserHandler<'_>

Creates a [users::UserHandler] for the specified user using the user ID

pub fn workflows( &self, owner: impl Into<String>, repo: impl Into<String>, ) -> WorkflowsHandler<'_>

Creates a workflows::WorkflowsHandler for the specified repository that allows you to access GitHub’s workflows API.

pub fn events(&self) -> EventsBuilder<'_>

Creates an events::EventsBuilder that allows you to access GitHub’s events API.

pub fn gists(&self) -> GistsHandler<'_>

Creates a gists::GistsHandler that allows you to access GitHub’s Gists API.

pub fn checks( &self, owner: impl Into<String>, repo: impl Into<String>, ) -> ChecksHandler<'_>

Creates a checks::ChecksHandler that allows to access the Checks API.

pub fn ratelimit(&self) -> RateLimitHandler<'_>

Creates a ratelimit::RateLimitHandler that returns the API rate limit.

pub fn hooks(&self, owner: impl Into<String>) -> HooksHandler<'_>

Creates a hooks::HooksHandler that returns the API hooks

impl Octocrab

GraphQL API.

pub async fn graphql<R: FromResponse>( &self, payload: &(impl Serialize + ?Sized), ) -> Result<R>

Sends a graphql query to GitHub, and deserialises the response from JSON.

let response: serde_json::Value = octocrab::instance()
    .graphql(&serde_json::json!({ "query": "{ viewer { login }}" }))
    .await?;

impl Octocrab

HTTP Methods

A collection of different of HTTP methods to use with Octocrab’s configuration (Authenication, etc.). All of the HTTP methods (get, post, etc.) perform some amount of pre-processing such as making relative urls absolute, and post processing such as mapping any potential GitHub errors into Err() variants, and deserializing the response body.

This isn’t always ideal when working with GitHub’s API and as such there are additional methods available prefixed with _ (e.g. _get, _post, etc.) that perform no pre or post processing and directly return the http::Response struct.

pub async fn post<P: Serialize + ?Sized, R: FromResponse>( &self, route: impl AsRef<str>, body: Option<&P>, ) -> Result<R>

Send a POST request to route with an optional body, returning the body of the response.

pub async fn _post<P: Serialize + ?Sized>( &self, uri: impl TryInto<Uri>, body: Option<&P>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Send a POST request with no additional pre/post-processing.

pub async fn get<R, A, P>(&self, route: A, parameters: Option<&P>) -> Result<R>
where A: AsRef<str>, P: Serialize + ?Sized, R: FromResponse,

Send a GET request to route with optional query parameters, returning the body of the response.

pub async fn _get( &self, uri: impl TryInto<Uri>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Send a GET request with no additional post-processing.

pub async fn body_to_string( &self, res: Response<BoxBody<Bytes, Error>>, ) -> Result<String>

pub async fn get_with_headers<R, A, P>( &self, route: A, parameters: Option<&P>, headers: Option<HeaderMap>, ) -> Result<R>
where A: AsRef<str>, P: Serialize + ?Sized, R: FromResponse,

Send a GET request to route with optional query parameters and headers, returning the body of the response.

pub async fn _get_with_headers( &self, uri: impl TryInto<Uri>, headers: Option<HeaderMap>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Send a GET request including option to set headers, with no additional post-processing.

pub async fn patch<R, A, B>(&self, route: A, body: Option<&B>) -> Result<R>
where A: AsRef<str>, B: Serialize + ?Sized, R: FromResponse,

Send a PATCH request to route with optional query parameters, returning the body of the response.

pub async fn _patch<B: Serialize + ?Sized>( &self, uri: impl TryInto<Uri>, body: Option<&B>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Send a PATCH request with no additional post-processing.

pub async fn put<R, A, B>(&self, route: A, body: Option<&B>) -> Result<R>
where A: AsRef<str>, B: Serialize + ?Sized, R: FromResponse,

Send a PUT request to route with optional query parameters, returning the body of the response.

pub async fn _put<B: Serialize + ?Sized>( &self, uri: impl TryInto<Uri>, body: Option<&B>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Send a PATCH request with no additional post-processing.

pub fn build_request<B: Serialize + ?Sized>( &self, builder: Builder, body: Option<&B>, ) -> Result<Request<OctoBody>>

pub async fn delete<R, A, B>(&self, route: A, body: Option<&B>) -> Result<R>
where A: AsRef<str>, B: Serialize + ?Sized, R: FromResponse,

Send a DELETE request to route with optional query body, returning the body of the response.

pub async fn _delete<B: Serialize + ?Sized>( &self, uri: impl TryInto<Uri>, body: Option<&B>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Send a DELETE request with no additional post-processing.

pub async fn send( &self, request: Request<OctoBody>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Send the given request to the underlying service

pub async fn execute( &self, request: Request<impl Into<OctoBody>>, ) -> Result<Response<BoxBody<Bytes, Error>>>

Execute the given request using octocrab’s Client.

pub async fn follow_location_to_data( &self, response: Response<BoxBody<Bytes, Error>>, ) -> Result<Response<BoxBody<Bytes, Error>>>

impl Octocrab

Utility Methods

pub async fn get_page<R: DeserializeOwned>( &self, uri: &Option<Uri>, ) -> Result<Option<Page<R>>>

A convenience method to get a page of results (if present).

pub async fn all_pages<R: DeserializeOwned>( &self, page: Page<R>, ) -> Result<Vec<R>>

A convenience method to get all the results starting at a given page.