Struct octocrab::OctocrabBuilder

pub struct OctocrabBuilder<Svc, Config, Auth, LayerReady> { /* private fields */ }

A builder struct for Octocrab, allowing you to configure the client, such as using GitHub previews, the github instance, authentication, etc.

let octocrab = octocrab::OctocrabBuilder::default()
    .add_preview("machine-man")
    .base_uri("https://github.example.com")?
    .build()?;

A builder struct for Octocrab. OctocrabBuilder can be extended with a custom config, see DefaultOctocrabBuilderConfig for an example

Implementations

impl OctocrabBuilder<NoSvc, NoConfig, NoAuth, NotLayerReady>

pub fn new_empty() -> Self

impl OctocrabBuilder<NoSvc, DefaultOctocrabBuilderConfig, NoAuth, NotLayerReady>

pub fn new() -> Self

impl<Config, Auth> OctocrabBuilder<NoSvc, Config, Auth, NotLayerReady>

pub fn with_service<Svc>( self, service: Svc, ) -> OctocrabBuilder<Svc, Config, Auth, LayerReady>

impl<Svc, Config, Auth, B> OctocrabBuilder<Svc, Config, Auth, LayerReady>
where Svc: Service<Request<OctoBody>, Response = Response<B>> + Send + 'static, Svc::Future: Send + 'static, Svc::Error: Into<BoxError>, B: Body<Data = Bytes> + Send + 'static, B::Error: Into<BoxError>,

pub fn with_executor( self, executor: Box<dyn Fn(Pin<Box<dyn Future<Output = ()>>>)>, ) -> OctocrabBuilder<Svc, Config, Auth, LayerReady>

impl<Svc, Config, Auth, B> OctocrabBuilder<Svc, Config, Auth, LayerReady>
where Svc: Service<Request<OctoBody>, Response = Response<B>> + Send + 'static, Svc::Future: Send + 'static, Svc::Error: Into<BoxError>, B: Body<Data = Bytes> + Send + 'static, B::Error: Into<BoxError>,

pub fn with_layer<L: Layer<Svc>>( self, layer: &L, ) -> OctocrabBuilder<L::Service, Config, Auth, LayerReady>

Add a Layer to the current Service stack.

impl<Svc, B, LayerState> OctocrabBuilder<Svc, NoConfig, AuthState, LayerState>
where Svc: Service<Request<OctoBody>, Response = Response<B>> + Send + 'static, Svc::Future: Send + 'static, Svc::Error: Into<BoxError>, B: Body<Data = Bytes> + Send + Sync + 'static, B::Error: Into<BoxError>,

pub fn build(self) -> Result<Octocrab, Infallible>

Build a [Client] instance with the current Service stack.

impl<Svc, Config, LayerState> OctocrabBuilder<Svc, Config, NoAuth, LayerState>

pub fn with_auth<Auth>( self, auth: Auth, ) -> OctocrabBuilder<Svc, Config, Auth, LayerState>

impl OctocrabBuilder<NoSvc, DefaultOctocrabBuilderConfig, NoAuth, NotLayerReady>

pub fn add_retry_config(self, retry_config: RetryConfig) -> Self

Set the retry configuration

pub fn set_connect_timeout(self, timeout: Option<Duration>) -> Self

Set the connect timeout.

pub fn set_read_timeout(self, timeout: Option<Duration>) -> Self

Set the read timeout.

pub fn set_write_timeout(self, timeout: Option<Duration>) -> Self

Set the write timeout.

pub fn add_preview(self, preview: &'static str) -> Self

Enable a GitHub preview.

pub fn add_header(self, key: HeaderName, value: String) -> Self

Add an additional header to include with every request.

pub fn personal_token<S: Into<SecretString>>(self, token: S) -> Self

Add a personal token to use for authentication.

pub fn app(self, app_id: AppId, key: EncodingKey) -> Self

Authenticate as a Github App. key: RSA private key in DER or PEM formats.

pub fn basic_auth(self, username: String, password: String) -> Self

Authenticate as a Basic Auth username and password

pub fn oauth(self, oauth: OAuth) -> Self

Authenticate with an OAuth token.

pub fn user_access_token<S: Into<SecretString>>(self, token: S) -> Self

Authenticate with a user access token.

pub fn base_uri(self, base_uri: impl TryInto<Uri>) -> Result<Self>

Set the base url for Octocrab.

pub fn upload_uri(self, upload_uri: impl TryInto<Uri>) -> Result<Self>

Set the base upload url for Octocrab.

pub fn set_connector_retry_service<S>( &self, connector: Client<S, OctoBody>, ) -> Retry<RetryConfig, Client<S, OctoBody>>

pub fn set_connect_timeout_service<T>( &self, connector: T, ) -> TimeoutConnector<T>
where T: Service<Uri> + Send, T::Response: Read + Write + Send + Unpin, T::Future: Send + 'static, T::Error: Into<BoxError>,

pub fn build(self) -> Result<Octocrab>

Build a [Client] instance with the current Service stack.